docker/cost-dashboard/requirements.txt
flask==3.1.0
gunicorn==23.0.0
google-cloud-storage==2.19.0
